% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caafidata-data.R
\docType{data}
\name{caafidata}
\alias{caafidata}
\title{CAAFI Data: Computer Aversion, Attitudes, and Familiarity Inventory}
\format{
A data frame with 794 rows and 30 variables.

\describe{
  \item{q1}{I enjoy using computers.}
  \item{q2}{Being able to use a computer is important to me.}
  \item{q3}{I keep up with the latest computer hardware.}
  \item{q4}{Computers are beneficial because they save people time.}
  \item{q5}{I like using word-processing programs.}
  \item{q6}{I feel like a fool when I am using a computer and others are around.}
  \item{q7}{I am smart enough to use a computer.}
  \item{q8}{I avoid using computers whenever possible.}
  \item{q9}{I do not understand how to use computer software (e.g., word-processing programs, spreadsheet programs, etc.).}
  \item{q10}{I feel that I understand how to use computer files, documents, and folders.}
  \item{q11}{I use a computer input device every day (e.g., a keyboard, a touch pad, a mouse).}
  \item{q12}{I can use a computer to successfully perform tasks.}
  \item{q13}{I can add new hardware to a computer.}
  \item{q14}{I enjoy reading computer magazines.}
  \item{q15}{When I use a computer, I am afraid that I will damage it.}
  \item{q16}{I enjoy connecting new computer accessories.}
  \item{q17}{I must have a reference manual or a help file to run computer software.}
  \item{q18}{E-mail is an easy way to communicate with people.}
  \item{q19}{I use e-mail every day.}
  \item{q20}{I am comfortable changing (installing/upgrading) computer software.}
  \item{q21}{I often read computer books.}
  \item{q22}{My friends often ask me computer-related questions.}
  \item{q23}{I often read computer magazines.}
  \item{q24}{Overall, I feel that I don't know how to use a computer.}
  \item{q25}{Computers are too scientific for me.}
  \item{q26}{When using a computer, I often lose data.}
  \item{q27}{I enjoy learning to use new software programs.}
  \item{q28}{I like to use computer input devices such as a keyboard, a touch pad, a mouse, etc.}
  \item{q29}{Using a computer is entertaining.}
  \item{q30}{I keep up with the latest computer software.}
}
}
\usage{
data(efa)
}
\description{
Study: This dataset has data collected on the computer
aversion, attitudes, and familiarity inventory.
}
\details{
The instructions were:
Below is a list of items describing many of the
thoughts and experiences that people have with computers.
After reading each statement, circle the number that best
describes how true or how false the statement is as it
applies to you at this time. If you have no opinion about
the item, circle ‘‘0”, but please use this option only if
it is absolutely necessary. Be sure to circle only one
number. Please do your best to respond to each item.

Scale -3 to 3: absolutely false, neutral, absolutely true

Computer Familiarity: Items 3, 13-14, 16, 20-23, 27, and 30.

Computer Attitudes: Items 1-2, 4-5, 8, 11, 18-19, and 28-29.

Computer Aversion: Items 6-7, 9-10, 12, 15, 17, and 24-26.
}
\keyword{datasets}
