% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dassdata-data.R
\docType{data}
\name{dassdata}
\alias{dassdata}
\title{DASS Data: Depression, Anxiety, and Stress Inventory}
\format{
A data frame with 794 rows and 30 variables.

\describe{
  \item{q1}{I found it hard to wind down.}
  \item{q2}{I was aware of dryness of my mouth.}
  \item{q3}{I couldn't seem to experience any positive feeling at all.}
  \item{q4}{I experienced breathing difficulty.}
  \item{q5}{I found it difficult to work up the initiative to do things.}
  \item{q6}{I tended to over-react to situations.}
  \item{q7}{I experienced trembling (eg, in the hands).}
  \item{q8}{I felt that I was using a lot of nervous energy.}
  \item{q9}{I was worried about situations in which I might panic and make a fool of myself.}
  \item{q10}{I felt that I had nothing to look forward to.}
  \item{q11}{I found myself getting agitated.}
  \item{q12}{I found it difficult to relax.}
  \item{q13}{I felt down-hearted and blue.}
  \item{q14}{I was intolerant of anything that kept me from getting on with what I was doing.}
  \item{q15}{I felt I was close to panic.}
  \item{q16}{I was unable to become enthusiastic about anything.}
  \item{q17}{I felt I wasn't worth much as a person.}
  \item{q18}{I felt that I was rather touchy.}
  \item{q19}{I was aware of the action of my heart in the absence of physical exertion.}
  \item{q20}{I felt scared without any good reason.}
  \item{q21}{I felt that life was meaningless.}
}
}
\usage{
data(dassdata)
}
\description{
Study: The DASS is a measurement scale that examines
the depression, anxiety, and stress of an individual.
}
\details{
The instructions were:
Please read each statement and select a number 0, 1,
2 or 3 that indicates how much the statement applied
to you over the past week. There are no right or wrong
answers. Do not spend too much time on any statement,
but please answer each question.

Scale 0-3: did not apply to me, applied to me to some
degree, applied to me to a considerable degree, or
applied to me very much

Depression: Questions 3, 5, 10, 13, 16, 17, 21

Anxiety: Questions 2, 4, 7, 9, 15, 19, 20

Stress: 1, 6, 8, 11, 12, 14, 18
}
\keyword{datasets}
