% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascreen-data.R
\docType{data}
\name{datascreen}
\alias{datascreen}
\title{Datascreening Practice Dataset}
\format{
A data frame with 797 rows and 12 variables,
see question descriptions above. The Participant_ID
column is the unique number for each participant.
}
\usage{
data(datascreen)
}
\description{
Study: This dataset includes a male body dissatisfaction scale with the following questions:
q1.	I think my body should be leaner
q2.	I am concerned that my stomach is too flabby
q3.	I feel dissatisfied with my overall body build
q4.	I think I have too much fat on my body
q5.	I think my abs are not thin enough
q6.	I feel satisfied with the size and shape of my body
q7.	Has eating sweets, cakes, or other high calorie food made you feel fat or weak?
q8.	Have you felt excessively large and rounded (i.e., fat)?
q9.	Have you felt ashamed of your body size or shape?
q10.	Has seeing your reflection (e.g., in a mirror or window) made you feel badly about your size or shape?
q11.	Have you been so worried about your body size or shape that you have been feeling that you ought to diet?
}
\keyword{datasets}
