% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtdata-data.R
\docType{data}
\name{dirtdata}
\alias{dirtdata}
\title{Dichotomous IRT Practice Data}
\format{
A data frame with 30 rows and 4 variables.
}
\usage{
data(dirtdata)
}
\description{
Study: This data represents the answers on an
Educational Psychology exam. A zero indicates that
the person missed the question, while one
indicates that the person got the question right.
}
\keyword{datasets}
