% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa-data.R
\docType{data}
\name{efa}
\alias{efa}
\title{Exploratory Factor Analysis Practice Dataset}
\format{
A data frame with 99 rows and 21 variables.

\describe{
  \item{o1}{Believe in the importance of art.}
  \item{o2}{Have a vivid imagination.}
  \item{o3}{Tend to vote for liberal political candidates.}
  \item{o4}{Carry the conversation to a higher level.}
  \item{o5}{Enjoy hearing new ideas.}
  \item{o6}{Enjoy thinking about things.}
  \item{o7}{Can say things beautifully.}
  \item{o8}{Enjoy wild flights of fantasy.}
  \item{o9}{Get excited by new ideas.}
  \item{o10}{Have a rich vocabulary.}
  \item{o11}{Am not interested in abstract ideas.}
  \item{o12}{Do not like art.}
  \item{o13}{Avoid philosophical discussions.}
  \item{o14}{Do not enjoy going to art museums.}
  \item{o15}{Tend to vote for conservative political candidates.}
  \item{o16}{Do not like poetry.}
  \item{o17}{Rarely look for a deeper meaning in things.}
  \item{o18}{Believe that too much tax money goes to support artists.}
  \item{o19}{Am not interested in theoretical discussions.}
  \item{o20}{Have difficulty understanding abstract ideas.}
  \item{condition}{a group condition each participant received}
}
}
\usage{
data(efa)
}
\description{
Study: This dataset has data on the Openness to
Experience scale collected as part of an undergraduate
honor's thesis project.
}
\details{
The instructions were:
Below are some phrases describing people's behaviors.
Please use the rating scale below to describe how
accurately each statement describes you. Describe
yourself as you generally are now, not as you wish to
be in the future. Describe yourself as you honestly
see yourself in relation to other people of your
gender and of roughly your same age.  Please read
each statement carefully, and then check the box
that corresponds to your response.

Scale: very inaccurate, moderately inaccurate,
neither inaccurate nor accurate, moderately
accurate, very accurate
}
\keyword{datasets}
