% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_context.R
\name{is_server_context}
\alias{is_server_context}
\title{Server Functions for learnr Tutorials}
\usage{
is_server_context(.envir)
}
\arguments{
\item{.envir}{Automatically grabs the environment
variable for your shiny session.}
}
\value{
Error messages if you incorrectly use
the functions.
}
\description{
These functions help check that you have put
together the tutorial correctly for the
student answers to print out at the end of the
tutorial
}
\keyword{answers}
\keyword{learnr,}
\keyword{shiny,}
\keyword{student}
