% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder_logic.R
\name{encoder_logic}
\alias{encoder_logic}
\title{Encoding Logic for learnr Tutorials}
\usage{
encoder_logic()
}
\value{
HTML output for the student tutorial
}
\description{
This function grabs the student answers from a learnr
tutorial and returns them as an HTML output for
printing to the tutorial screen.
}
\examples{

# Be sure to put this into a server-context chunk.
#```{r context="server"}
#encoder_logic()
#```
}
\keyword{answers}
\keyword{learnr,}
\keyword{shiny,}
\keyword{student}
